requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.4.1-cp37-cp37m-win32.whl
pytest-allure-adaptor